requirements.txt
flask
pandas
joblib
scikit-learn
pymongo
python-dotenv
pyjwt
werkzeug